file routes, and rename middleware auth file to setup file
